# [b200] fuse allreduce+add_rmsnorm in prepare_attention + mlp module