--- a/python/sglang/srt/layers/communicator.py
+++ b/python/sglang/srt/layers/communicator.py
@@ -36,6 +36,7 @@ from sglang.srt.layers.utils import is_sm100_supported
 from sglang.srt.managers.schedule_batch import global_server_args_dict
 from sglang.srt.model_executor.forward_batch_info import ForwardBatch
 from sglang.srt.utils import is_cuda, is_flashinfer_available
+from sglang.srt.layers.flashinfer_comm_fusion import flashinfer_allreduce_add_rmsnorm
 
 _is_flashinfer_available = is_flashinfer_available()
 _is_sm100_supported = is_cuda() and is_sm100_supported()
@@ -188,8 +189,7 @@ class LayerCommunicator:
             residual = hidden_states
         else:
             if residual is not None and hasattr(hidden_states, '_sglang_needs_allreduce_fusion') and hidden_states._sglang_needs_allreduce_fusion:
-                fused_result = self._try_allreduce_add_rmsnorm_fusion(hidden_states, residual)
-                assert fused_result is not None, "Fused result should not be None"
+                fused_result = self._allreduce_add_rmsnorm_fusion(hidden_states, residual)
                 delattr(hidden_states, '_sglang_needs_allreduce_fusion')
                 hidden_states, residual = fused_result    
             else:
@@ -207,16 +207,8 @@ class LayerCommunicator:
 
         return hidden_states, residual
 
-    def _try_allreduce_add_rmsnorm_fusion(self, hidden_states: torch.Tensor, residual: torch.Tensor):
-        from sglang.srt.layers.flashinfer_comm_fusion import flashinfer_allreduce_add_rmsnorm
-        from sglang.srt.distributed import get_tensor_model_parallel_world_size
-        
-        if get_tensor_model_parallel_world_size() <= 1:
-            return None
-            
-        if residual is None:
-            return None
-            
+    def _allreduce_add_rmsnorm_fusion(self, hidden_states: torch.Tensor, residual: torch.Tensor):
+
         norm_out, residual_out = flashinfer_allreduce_add_rmsnorm(
             input_tensor=hidden_states,
             residual=residual,
@@ -224,10 +216,7 @@ class LayerCommunicator:
             eps=self.input_layernorm.variance_epsilon,
         )
         
-        if norm_out is not None:
-            return norm_out, residual_out
-        else:
-            return None
+        return norm_out, residual_out
 
     def prepare_mlp(
         self,