<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## b200 acc

main:

```shell
➜  sglang python3 benchmark/gsm8k/bench_sglang.py --num-questions 2000 --parallel 2000 --num-shots 8
100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1319/1319 [03:03<00:00,  7.17it/s]
Accuracy: 0.901
Invalid: 0.000
Latency: 184.588 s
Output throughput: 740.189 token/s
```

pr:

```shell
➜  sglang python3 benchmark/gsm8k/bench_sglang.py --num-questions 2000 --parallel 2000 --num-shots 8
100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1319/1319 [03:03<00:00,  7.17it/s]
Accuracy: 0.901
Invalid: 0.000
Latency: 184.588 s
Output throughput: 740.189 token/s
```


## profile

main:

![图片](https://github.com/user-attachments/assets/65c5e747-2417-4bd6-9622-690159fe3d7e)

pr:

![图片](https://github.com/user-attachments/assets/7426f5bb-615e-4e02-811d-a7c92242213c)

allreduce+add_rmsnorm+dsv3_fused_a_gemm: 22us->8us.

## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.